requirements.txt
replicate
flask==2.3.3
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
python-multipart==0.0.6
Werkzeug==2.3.7